content の pre/next ID 取得処理は FE 側で行うように変更
cf. #226
