+++ components/publish/menus/detail/list.vue
