+++ pages/services/[id].vue
@@ -1,17 +1,7 @@
-<script setup lang="ts">
-import type ServicesList from '@/components/publish/services/list.vue'
-const servicesListRef = ref<InstanceType<typeof ServicesList> | null>(null)
-</script>
-
 <template>
   <div id="service-detail-top">
     <PublishLayoutNavBreadCrumb class="nav-bread-crumbs" />
-    <PublishServicesDetail @update:data="servicesListRef?.onGetList" />
-    <PublishServicesList
-      ref="servicesListRef"
-      no-caption
-      class="service-list-gap"
-    />
+    <PublishServicesDetail />
   </div>
 </template>
 
@@ -26,9 +16,5 @@ const servicesListRef = ref<InstanceType<typeof ServicesList> | null>(null)
   @media only screen and (max-width: $grid-breakpoint-md) {
     margin-bottom: 0;
   }
-}
-
-.service-list-gap {
-  margin-top: 6rem;
 }
 </style>
